requirements.txt
numpy>=1.24.0
pandas>=2.1.0
scikit-learn>=1.4.2
scipy>=1.10.0
matplotlib>=3.7.0
seaborn>=0.13.0
PyYAML>=6.0
joblib>=1.3.0

